+++ src/components/app/details/triggerView/ciMaterial.tsx
@@ -106,8 +106,6 @@ export class CIMaterial extends Component<CIMaterialProps> {
             webhookPayloads={this.props.webhookPayloads}
             isWebhookPayloadLoading={this.props.isWebhookPayloadLoading}
             workflowId={this.props.workflowId}
-            onClickWebhookTimeStamp={this.props.onClickWebhookTimeStamp}
-            webhhookTimeStampOrder={this.props.webhhookTimeStampOrder}
           />
         </div>
         {this.props.showWebhookModal ? null : this.renderMaterialStartBuild(context, canTrigger)}
