+++ src/components/app/details/triggerView/ciWebhook.service.ts
@@ -1,7 +1,7 @@
 import { get } from '../../../../services/api';
 
 export function getCIWebhookRes(pipelineMaterialId, timeStampOrder): Promise<any> {
-    const URL = `app/ci-pipeline/webhook-payload/${pipelineMaterialId}?limit=1000&offset=0&timeSort=DSC`;
+    const URL = `app/ci-pipeline/webhook-payload/${pipelineMaterialId}?limit=1000&offset=0&timeSort=DESC`;
     return get(URL);
 }
